ci: p100a: temporarily disable p100a tests due to jtag reset issues
Temporarily disable P100A tests due to lack of JTAG SRST signal.

This affects P150A and other platforms as well. P100 is the only
platform not affected.

The tests will be re-enabled when issue 101 is fixed. For any bugfixes
targeting P100A+, a "Testing Done" section will need to be included
with the PR description, as we are in feature freeze / stabilization.

https://github.com/tenstorrent/tt-zephyr-platforms/issues/101

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>