ci: hardware: update hw ci to define matrix with check-runners.py
Update hardware CI workflows to query which runners are available from

a) the organization, and / or
b) the repository

to evaluate whether to run.

The way that this works, is that .github/runners.yml contains a
list of board (revision) / label combinations.

./scripts/check-runners.py loads the list of possible runners from that
file, and filters-out runners that are not present after querying status
from the GitHub REST API.

If no runners are present, then the hardware workflows will not run.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>