scripts: recovery: program spi flash via stm32+pyocd
Initial attempt at creating a programmatic way of recovering boards via
python and pyocd.

This approach uses any SWD-compatible debugger to connect to the stm32.

First, SWD sequences are sent to the STM32 to

* reset and halt the stm32 core

Then, a Cortex-M FLM (flash management algorithm) is copied into SRAM.
The flash algorithm is a set of Thumb2 instructions that implement very
specific routines, specific to the stm32 in order to:

* set up pinmux for SPI SCK, MOSI, MOSI, and #SS
* program the peripheral clock registers
* initialize the SPI transceiver

The Program Counter value that must be set to execute these actions are
labeled "pc_init"

There are additional sequences of Thumb2 instructions. Similarly, the
Program Counter values that must be set to execute each of those are
"pc_program_page", "pc_erase_sector", and "pc_eraseAll".

Additionally, there is an Program Counter value that must be set called
"pc_uninit" that will

* deinitialize the SPI transceiver
* disable the peripheral clocks
* put pins into high-impedence state

This is a WIP, and just barely started to function.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>