doc: normalize doc builds around the doc directory
Previously, the `boards` directory of the top-level repo was referenced
by the sphinx build and was difficult to include with the desired
documentation and a workaround was done by changing the sphinx build
directory in the Makefile.

Undue the workaround and simply use a symlink in GitHub Actions.

This is required because symlinks trip check_compliance.py.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>