manifest: switch to zephyr v4.3.0-pre
Update to Zephry v4.2.0-pre.

This change updates to a relatively stable version of Zephyr
that is about 1 month before the feature freeze of v4.3.0-rc1.

Updating to this version allows us to drop a few patches immediately.

As we approach the v4.3.0-rc1 release candidate (and feature
freeze), we should aim to reduce the number of patches to zero
(or a more manageable, smaller number).

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>