ci: consolodate build stesp in build-fw workflow
Rather than building firmware in the release workflow, build it in the
build-fw workflow.

Additionally, de-duplicate the number of jobs that happen in parallel.

Most of the build steps can share a common checkout.

The theory is, that by using ccache, we should be able to blow-away any
gains made by parallelizing the build so explosively.

The main benefit here is that CI is more maintainable, and theere is less
copy-paste garbage literally everywhere.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>