ci: build-fw: build preflash.ihex with a cmd rather than a whole job
The build-preflash.sh script was kind of overkill for what amounted to
running a single command.

```shell
./scripts/tt_boot_fs.py \
            mkfs \
            "$YAML_PATH" \
            preflash.ihex \
            --build-dir "$TTZP/build-${{ matrix.board }}" \
            --hex
```

Remove the script and use the single necessary command instead.

Remove the additional build job that duplicates all of the prepare-zephyr
reuseable workflow (including fetching sources, patching the Zephyr tree,
etc) along with the duplicate p100a build. We already do all of that in the
`build` job, so let's just reuse the already-built firmware directory.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>