ci: build-fw: move flm build job to 'build'
Move the "Build Flash Loader Module" job to a single step inside of the
'build' job to de-duplicate checking out all of the Zephyr sources.

Since FLM depends on the stm32 HAL being present, we do still require
a Zephyr checkout (via prepare-zephyr). Otherwise, it would be preferable
to keep it in a separate job with far fewer setup steps.

Signed-off-by: Chris Friedt <cfriedt@tenstorrent.com>